+++ kernel/arch/dreamcast/kernel/irq.c
@@ -2,11 +2,16 @@
 
    arch/dreamcast/kernel/irq.c
    Copyright (C) 2000-2001 Megan Potter
+   Copyright (C) 2024 Paul Cercueil
+   Copyright (C) 2024 Falco Girgis
+   Copyright (C) 2024 Andy Barajas
 */
 
 /* This module contains low-level handling for IRQs and related exceptions. */
 
 #include <string.h>
+#include <strings.h>
+#include <assert.h>
 #include <stdio.h>
 #include <arch/arch.h>
 #include <arch/types.h>
@@ -17,74 +22,177 @@
 #include <kos/thread.h>
 #include <kos/library.h>
 
+/* Macros for accessing related registers. */
+#define TRA    ( *((volatile uint32_t *)(0xff000020)) ) /* TRAPA Exception Register */
+#define EXPEVT ( *((volatile uint32_t *)(0xff000024)) ) /* Exception Event Register */
+#define INTEVT ( *((volatile uint32_t *)(0xff000028)) ) /* Interrupt Event Register */
+
+/* IRQ handler closure */
 struct irq_cb {
     irq_handler hdl;
-    void *data;
+    void       *data;
 };
 
-/* Exception table -- this table matches (EXPEVT>>4) to a function pointer.
-   If the pointer is null, then nothing happens. Otherwise, the function will
-   handle the exception. */
-static struct irq_cb irq_handlers[0x100];
-static struct irq_cb trapa_handlers[0x100];
+/* TRAPA handler closure */
+struct trapa_cb {
+    trapa_handler hdl;
+    void         *data;
+};
 
-/* Global exception handler -- hook this if you want to get each and every
-   exception; you might get more than you bargained for, but it can be useful. */
-static irq_handler irq_hnd_global;
-static void *irq_hnd_global_data;
+/* Linked list of IRQ states, one is pushed onto the stack
+   every time the top-level ISR is entered. */
+struct irq_state {            // SIZE
+    bool           handled;   // 1 byte  /* mov.b only has 15-byte displacement */
+    uint8_t        code;      // 1 byte
+    uint16_t       evt;       // 2 bytes
+    volatile struct
+        irq_state *previous;  // 4 bytes
+};                            // 8 BYTES TOTAL
 
+/* Individual exception handlers */
+static struct irq_cb   irq_handlers[0x100];
+/* TRAPA exception handlers. */
+static struct trapa_cb trapa_handlers[0x100];
+/* Global exception handler */
+static struct irq_cb   global_irq_handler;
 /* Default IRQ context location */
-static irq_context_t irq_context_default;
+static irq_context_t   irq_context_default;
+/* Current IRQ state linked list pointer */
+static volatile struct irq_state *        /* Points to most recent IRQ state */
+       volatile        irq_state_current; /* Volatile pointer to volatile struct */
+
+/* Called when entiringan IRQ to push its state onto the stack as current. */
+inline static void irq_state_push(struct irq_state *current) {
+    current->previous = irq_state_current;
+    irq_state_current = current;
+}
+
+/* Called when exiting an IRQ to pop its state from the stack. */
+inline static void irq_state_pop(void) {
+    assert(irq_state_current);
+    irq_state_current = irq_state_current->previous;
+}
+
+/* Called to grab the IRQ state at a particular stack level. */
+inline static volatile struct irq_state *irq_state_n(size_t level) {
+    volatile struct irq_state *state = irq_state_current;
+    
+    for(size_t depth = 0; depth < level; ++depth) {
+        if(!state) break;
+        state = state->previous;
+    }
+
+    return state;
+}
+
+/* How deeply nested in IRQ calls are we? */
+size_t irq_int_depth(void) {
+    size_t depth = 0;
+    const irq_mask_t imask = irq_disable();
+
+    const volatile struct irq_state *state = irq_state_current;
+    
+    while(state) {
+        ++depth;
+        state = state->previous;
+    }
+
+    irq_restore(imask);
+
+    return depth;
+}
 
 /* Are we inside an interrupt? */
-static int inside_int;
-int irq_inside_int(void) {
-    return inside_int;
+bool irq_inside_int(void) {
+    return !!irq_state_current;
+}
+
+/* What's the active IRQ? */
+irq_t irq_active_int(size_t level) {
+    const irq_mask_t imask = irq_disable();
+
+    const volatile struct irq_state *state = irq_state_n(level);
+    const irq_t irq = state? state->evt : 0;
+
+    irq_restore(imask);
+    return irq;
+}
+
+/* Have we handled the active interrupt? */
+bool irq_handled_int(size_t level) {
+    const irq_mask_t imask = irq_disable();
+
+    const volatile struct irq_state *state = irq_state_n(level);
+    assert(state);
+    const bool handled = state->handled;
+    
+    irq_restore(imask);
+    return handled;
+}
+
+/* Set whether we've handled the active interrupt or not. */
+void irq_handle_int(bool handled) {
+    assert(irq_state_current);
+    irq_state_current->handled = handled;
 }
 
 /* Set a handler, or remove a handler */
 int irq_set_handler(irq_t code, irq_handler hnd, void *data) {
     /* Make sure they don't do something crackheaded */
-    if(code >= 0x1000 || (code & 0x000f)) return -1;
+    if(code >= 0x1000 || (code & 0x000f)) 
+        return -1;
 
-    code = code >> 4;
+    code >>= 4;
     irq_handlers[code] = (struct irq_cb){ hnd, data };
 
     return 0;
 }
 
 /* Get the address of the current handler */
-irq_handler irq_get_handler(irq_t code) {
+irq_handler irq_get_handler(irq_t code, void **data) {
     /* Make sure they don't do something crackheaded */
-    if(code >= 0x1000 || (code & 0x000f)) return NULL;
+    if(code >= 0x1000 || (code & 0x000f))
+        return NULL;
+    
+    code >>= 4;
 
-    code = code >> 4;
+    if(data)
+        *data = irq_handlers[code].data;
+
     return irq_handlers[code].hdl;
 }
 
 /* Set a global handler */
 int irq_set_global_handler(irq_handler hnd, void *data) {
-    irq_hnd_global = hnd;
-    irq_hnd_global_data = data;
+    global_irq_handler.hdl = hnd;
+    global_irq_handler.data = data;
     return 0;
 }
 
 /* Get the global exception handler */
-irq_handler irq_get_global_handler(void) {
-    return irq_hnd_global;
+irq_handler irq_get_global_handler(void **data) {
+    if(data)
+        *data = global_irq_handler.data;
+
+    return global_irq_handler.hdl;
 }
 
 /* Set or remove a trapa handler */
-int trapa_set_handler(irq_t code, irq_handler hnd, void *data) {
-    if(code > 0xff) return -1;
-
-    trapa_handlers[code] = (struct irq_cb){ hnd, data };
-
+int trapa_set_handler(trapa_t code, trapa_handler hnd, void *data) {
+    trapa_handlers[code] = (struct trapa_cb){ hnd, data };
     return 0;
 }
 
+/* Get a particular trapa handler */
+trapa_handler trapa_get_handler(trapa_t code, void **data) {
+    if(data)
+        *data = trapa_handlers[code].data;
+
+    return trapa_handlers[code].hdl;
+}
+
 /* Get a string description of the exception */
-static char *irq_exception_string(int evt) {
+static char *irq_exception_string(irq_t evt) {
     switch(evt) {
         case EXC_ILLEGAL_INSTR:
             return "Illegal instruction";
@@ -121,14 +229,14 @@ static char *irq_exception_string(int evt) {
 
 /* Print a kernel panic reg dump */
 extern irq_context_t *irq_srt_addr;
-static void irq_dump_regs(int code, int evt) {
+static void irq_dump_regs(int code, irq_t evt) {
     uint32_t fp;
     uint32_t *regs = irq_srt_addr->r;
     bool valid_pc;
     bool valid_pr;
 
     dbglog(DBG_DEAD, "Unhandled exception: PC %08lx, code %d, evt %04x\n",
-           irq_srt_addr->pc, code, (uint16)evt);
+           irq_srt_addr->pc, code, (uint16_t)evt);
     dbglog(DBG_DEAD, " R0-R7: %08lx %08lx %08lx %08lx %08lx %08lx %08lx %08lx\n",
            regs[0], regs[1], regs[2], regs[3], regs[4], regs[5], regs[6], regs[7]);
     dbglog(DBG_DEAD, " R8-R15: %08lx %08lx %08lx %08lx %08lx %08lx %08lx %08lx\n",
@@ -179,107 +287,97 @@ static void irq_dump_regs(int code, int evt) {
 /* The C-level routine that processes context switching and other
    types of interrupts. NOTE: We are running on the stack of the process
    that was interrupted! */
-volatile uint32_t jiffies = 0;
 void irq_handle_exception(int code) {
-    /* Get the exception code */
-    /* volatile uint32_t *tra = (uint32_t*)0xff000020; */
-    volatile uint32_t *expevt = (uint32_t *)0xff000024;
-    volatile uint32_t *intevt = (uint32_t *)0xff000028;
+    struct irq_state irq_state = {
+        .code = code
+    };
     const struct irq_cb *hnd;
-    uint32_t evt = 0;
-    int handled = 0;
 
-    /* If it's a code 0, well, we shouldn't be here. */
-    if(code == 0) arch_panic("spurious RESET exception");
+    irq_state_push(&irq_state);
+    
+    switch(code) {
+        /* If it's a code 0 (or anything else), well, we shouldn't be here. */
+        case 0:
+        default:
+            arch_panic("Spurious RESET exception!");
+            break;
 
-    /* If it's a code 1 or 2, grab the event from expevt. */
-    if(code == 1 || code == 2) evt = *expevt;
+        /* If it's a code 1 or 2, grab the event from expevt. */
+        case 1:
+        case 2:
+            irq_state.evt = EXPEVT;
+            break;
 
-    /* If it's a code 3, grab the event from intevt. */
-    if(code == 3) evt = *intevt;
+        /* If it's a code 3, grab the event from intevt. */
+        case 3:
+            irq_state.evt = INTEVT;
+    }
 
-    if(inside_int) {
+    /* Check for double exception fault: special case since we do not
+       currently support nesting of exceptions. */
+    if(__unlikely(irq_state.previous)) {
         hnd = &irq_handlers[EXC_DOUBLE_FAULT >> 4];
-        if(hnd->hdl != NULL)
+
+        if(hnd->hdl)
             hnd->hdl(EXC_DOUBLE_FAULT, irq_srt_addr, hnd->data);
-        else
-            irq_dump_regs(code, evt);
-
-        thd_pslist(dbgio_printf);
-        // library_print_list(dbgio_printf);
-        arch_panic("double fault");
-    }
-
-    /* Reveal this info about the int to inside_int for better 
-       diagnostics returns if we try to do something in the int. */
-    inside_int = ((code&0xf)<<16) | (evt&0xffff);
-
-    /* If there's a global handler, call it */
-    if(irq_hnd_global) {
-        irq_hnd_global(evt, irq_srt_addr, irq_hnd_global_data);
-        handled = 1;
-    }
-
-    /* dbgio_printf("got int %04x %04x\n", code, evt); */
-
-    /* If it's a timer interrupt, clear the status */
-    if(evt >= EXC_TMU0_TUNI0 && evt <= EXC_TMU2_TUNI2) {
-        if(evt == EXC_TMU0_TUNI0) {
-            timer_clear(TMU0);
-        }
-        else if(evt == EXC_TMU1_TUNI1) {
-            timer_clear(TMU1);
-        }
-        else {
-            timer_clear(TMU2);
-        }
-
-        handled = 1;
-    }
-
-    /* If there's a handler, call it */
-    {
-        hnd = &irq_handlers[evt >> 4];
-        if(hnd->hdl != NULL) {
-            hnd->hdl(evt, irq_srt_addr, hnd->data);
-            handled = 1;
+        
+        /* Panic if it went unhandled. */
+        if(!irq_state.handled) {
+            irq_dump_regs(code, irq_state.evt);
+            arch_panic("double fault");
         }
     }
 
-    if(!handled) {
+    /* If there's a global handler, it goes first */
+    if(__unlikely(global_irq_handler.hdl))
+        global_irq_handler.hdl(irq_state.evt, irq_srt_addr, global_irq_handler.data);
+
+    /* If the global handler didn't handle the exception, pass
+       it on to the individual handlers */
+    if(__likely(!irq_state.handled)) {
+        hnd = &irq_handlers[irq_state.evt >> 4];
+        
+        if(__likely(hnd->hdl)) {
+            /* Individual handlers accept by default. */
+            irq_state.handled = true;
+            hnd->hdl(irq_state.evt, irq_srt_addr, hnd->data);
+        }
+    }
+
+    /* If an individual handler didn't handle the exception,
+       pass it on to the unhandled exception handler. */
+    if(__unlikely(!irq_state.handled)) {
         hnd = &irq_handlers[EXC_UNHANDLED_EXC >> 4];
-        if(hnd->hdl != NULL)
-            hnd->hdl(evt, irq_srt_addr, hnd->data);
-        else
-            irq_dump_regs(code, evt);
 
-        arch_panic("unhandled IRQ/Exception");
+        if(hnd->hdl)
+            hnd->hdl(irq_state.evt, irq_srt_addr, hnd->data);
+        
+        /* Panic if nothing handled it. */
+        if(!irq_state.handled) {
+            irq_dump_regs(code, irq_state.evt);
+            arch_panic("unhandled IRQ/Exception");
+        }
     }
-
-    /* dbgio_printf("returning from int\n"); */
 
     irq_disable();
-    inside_int = 0;
+    irq_state_pop();
 }
 
 void irq_handle_trapa(irq_t code, irq_context_t *context, void *data) {
-    /* Get the exception code */
-    volatile uint32_t *tra = (uint32_t *)0xff000020;
     const struct irq_cb *hnd, *handlers = data;
     uint32_t vec;
 
     (void)code;
 
     /* Get the trapa vector */
-    vec = (*tra) >> 2;
+    vec = TRA >> 2;
 
     /* Check for handler and call if present */
     hnd = &handlers[vec];
 
-    if(hnd->hdl != NULL)
+    if(hnd->hdl)
         hnd->hdl(vec, context, hnd->data);
 }
-
 
 extern void irq_vma_table(void);
 
@@ -302,28 +400,12 @@ irq_context_t *irq_get_context(void) {
    or user mode. The given parameters will be passed to the called routine (up
    to the architecture maximum). */
 void irq_create_context(irq_context_t *context, uint32_t stkpntr,
-                        uint32_t routine, uint32_t *args, int usermode) {
-    int i;
-
-    /* Clear out user and FP regs */
-    for(i = 0; i < 16; i++) {
-        context->r[i] = 0;
-        context->fr[i] = 0;
-        context->frbank[i] = 0;
-    }
-
-    /* Set misc system regs */
-    context->gbr = context->mach = context->macl = 0;
-    context->vbr = 0;   /* This is not relevant because the
-                   context switcher doesn't touch it */
-
-    /* Set default floating point control regs */
-    context->fpscr = 0;
-    context->fpul = 0;
+                        uint32_t routine, const uint32_t *args, bool usermode) {
+    /* Clear out all registers. */
+    bzero(context, sizeof(irq_context_t));
 
     /* Setup the program frame */
     context->pc = (uint32_t)routine;
-    context->pr = 0;
     context->sr = 0x40000000;   /* note: need to handle IMASK */
     context->r[15] = stkpntr;
     context->r[14] = 0xffffffff;
@@ -346,6 +428,7 @@ static void irq_def_timer(irq_t src, irq_context_t *context, void *data) {
     (void)src;
     (void)context;
     (void)data;
+    timer_clear((int)data);
 }
 
 /* Default FPU exception handler (can't seem to turn these off) */
@@ -359,10 +442,12 @@ static void irq_def_fpu(irq_t src, irq_context_t *context, void *data) {
 static uint32_t pre_sr, pre_vbr;
 
 /* Have we been initialized? */
-static int initted = 0;
+static bool initted = false;
 
 /* Init routine */
 int irq_init(void) {
+    assert(!initted);
+
     /* Save SR and VBR */
     __asm__("stc    sr,r0\n"
             "mov.l  r0,%0" : : "m"(pre_sr));
@@ -373,16 +458,17 @@ int irq_init(void) {
     irq_disable();
 
     /* Blank the exception handler tables */
-    memset(irq_handlers, 0, sizeof(irq_handlers));
-    memset(trapa_handlers, 0, sizeof(trapa_handlers));
-    irq_hnd_global = NULL;
-    irq_hnd_global_data = NULL;
+    bzero(irq_handlers,        sizeof(irq_handlers));
+    bzero(trapa_handlers,      sizeof(trapa_handlers));
+    bzero(&global_irq_handler, sizeof(global_irq_handler));
 
     /* Default to not in an interrupt */
-    inside_int = 0;
+    irq_state_current = NULL;
 
-    /* Set a default timer handler */
-    irq_set_handler(TIMER_IRQ, irq_def_timer, NULL);
+    /* Set default timer handlers */
+    irq_set_handler(EXC_TMU0_TUNI0, irq_def_timer, (void *)0);
+    irq_set_handler(EXC_TMU1_TUNI1, irq_def_timer, (void *)1);
+    irq_set_handler(EXC_TMU2_TUNI2, irq_def_timer, (void *)2);
 
     /* Set a trapa handler */
     irq_set_handler(EXC_TRAPA, irq_handle_trapa, trapa_handlers);
@@ -406,7 +492,7 @@ int irq_init(void) {
             "	.long _irq_vma_table\n"
             "_after_vbr:\n");
 
-    initted = 1;
+    initted = true;
 
     return 0;
 }
@@ -421,5 +507,5 @@ void irq_shutdown(void) {
     __asm__("mov.l  %0,r0\n"
             "ldc    r0,vbr" : : "m"(pre_vbr));
 
-    initted = 0;
+    initted = false;
 }